==> 02_script/portugues/04_rendimento_medio.qmd <==
---
title: "Untitled"
format: html
editor: visual
---

## Salário médio de profissionais em saúde bocal.

O presente trabalho estima o salário médio das seguintes categorias de profissionais em saúde bocal utilizando dados da Pnad Contínua do segundo trimestre de 2024:

A\) Dentista (2261)\
B) Dentistas Auxiliares e Ajudantes de Odontologia (3251)

Bibliotecas utilizadas:

```{r}
library(PNADcIBGE)
library(survey)
library(writexl)
library(tidyverse)
library(readxl)
library(srvyr)
library(openxlsx)
```

Primeiramente, carregaremos os dados da Pnad Contínua do primeiro trimestre de 2024 que será utilizado tanto para estimar o salário médio de médicos quanto de auxiliares e ajudantes de odontologia.

```{r}

# Carregando a base da Pnad Contínua
base_pnadc <- get_pnadc(2024, 3, 
                        vars=c("VD4016","V4010"))

#HABILITANDO PARA O PACOTE SRVYR
base_pnad012024 <- as_survey(base_pnadc) 
  
# Calculando a nível Brasil
rend_dentista_br <- base_pnad012024 |> 
    filter(V4010 == "2261") |> 
    group_by(V4010) |> 
    summarise(rend_dentista = round(mean(VD4016, 
                                         na.rm = TRUE),
                                    2)) |>
    select(-ends_with("_se"))


# Calculando a média de dentistas por Estado  
  rend_dentista <- base_pnad012024 |> 
    filter(V4010 == "2261") |> 
    group_by(UF) |> 
    summarise(rend_dentista = round(mean(VD4016, 
                                         na.rm = TRUE),
                                    2)) |>
    select(-ends_with("_se"))
  
# Calculando a média de auxiliares e ajudantes de odontologia por Estado  
  rend_auxodonto <- base_pnad012024 |> 
    filter(V4010 == "3251") |> 
    group_by(UF) |> 
    summarise(rend_aux = round(mean(VD4016, na.rm = TRUE),2)) |> 
    select(-ends_with("_se"))
  
#Juntando as duas categorias em uma mesma base
rend_saudebocal <- rend_dentista |> 
  left_join(rend_auxodonto, by = c("UF" = "UF"))


```

Em seguida, iremos renomear a coluna UF de forma a ter apenas as suas respectivas siglas.

```{r}

# Tabela de correspondência
uf_siglas <- tibble::tibble(
  estado = c("Acre", "Alagoas", "Amapá", 
             "Amazonas", "Bahia", "Ceará", 
             "Distrito Federal", "Espírito Santo", 
             "Goiás", "Maranhão", "Mato Grosso", 
             "Mato Grosso do Sul", "Minas Gerais", 
             "Pará", "Paraíba", "Paraná", "Pernambuco", 
             "Piauí", "Rio de Janeiro", "Rio Grande do Norte", 
             "Rio Grande do Sul", "Rondônia", "Roraima", 
             "Santa Catarina","São Paulo", "Sergipe",
             "Tocantins"),
  sigla = c("AC", "AL", "AP", "AM", "BA", "CE", "DF", "ES", 
            "GO", "MA", "MT", "MS", "MG", "PA", "PB", "PR", 
            "PE", "PI", "RJ", "RN", "RS", "RO", "RR", "SC", 
            "SP", "SE", "TO")
)

# Realizar o join para renomear os estados pela sigla
rend_saudebocal <- rend_saudebocal %>%
  left_join(uf_siglas, by = c("UF" = "estado")) %>%
  select(-UF) %>%
  rename(UF = sigla) |> 
  select(UF, everything())

#exportando o arquivo em excel
write.xlsx(rend_saudebocal, file = "C:/Users/alefs/OneDrive/Documentos/saude_bucal/01_dados/rendimento_saudebocal.xlsx")
```
